Add the Psychedelic example.
Psychedelic is my attempt to re-create a display observed at a childrens museum
in which peoples outlines were projected in ~6 lagged frames of psychedelic
colors.  The code is a little hackish entirely because the code is a little
hackish, but I'll blame the fact that this is the ~4th base on which Psychedelic
was written and the porting jobs simply have not been stellar (ffmpeg-light,
then directly on freenect, then friday-freenect, and now on mellow).